+++ ui/templates/all.go
@@ -1,0 +1,9 @@
+package templates
+
+// GetTemplateList returns a list of all templates
+func GetTemplateList() []string {
+	var templates []string
+	templates = append(templates, "AppOverviewTemplate")
+
+	return templates
+}
